Rewrite the entire library
Some of these changes are superficial, but there's some significant
ones:

- Add writing ability
- Simplify the test cases to make adding new test cases easier

There is still no intelligent VLR handling — we read and write them, but
we don't learn anything from them.

There is also a failing test w/ an extrabytes roundrip. I need to push
out a new release for another project, so I'm going to commit this with
that test ignored, but that needs to be fixed.