Convert file configuration to `Builder`
It will be necessary to do some gatekeeping between user configuration
and the actual header, e.g. reconciling the vlrs with the extra bytes on
the point format. This patch adds a `Builder` structure that takes over
all configuration responsibilities, and makes `Header` read-only. This
lets us ensure that `Header`s are set up correctly.

The motivation here came from the workflow in a `Reader`, realizing that
we needed to build up the metadata incrementally before processing it in
one go. The `Builder` was useful in `Reader::new`, and it then became
clear that it should be the API method for creating headers as well.