Add ReaderOption struct
This new structs allows users to select whether they want to use
parallism when reading a LAZ file. By default, when the laz-parallel
feature is enabled, parallelism will be used (as it was before)

A new Reader::with_options is added to create a Reader with non-default
options.

Normally there is no API break